requirements.txt
streamlit
opencv-python-headless
mediapipe
numpy
pandas
groq
python-dotenv
scipy
dtaidistance
scikit-learn